Fixed error when running server. Reorganized database folders into new directory. Created comment resolver
