+++ database/models/index.js
@@ -29,7 +29,7 @@ fs.readdirSync(__dirname)
     file.indexOf('.') !== 0 && file !== basename && file.slice(-3) === '.js'
 ))
   .forEach((file) => {
-    const model = sequelize.import(path.join(__dirname, file));
+    const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes)
     db[model.name] = model;
   });
